refactor props for bulk granule components

+++ app/src/js/components/Granules/bulk.js
@@ -13,6 +13,16 @@ import {
 import BulkOperationsModal from './bulk-granule-operations';
 import BulkDeleteModal from './bulk-granule-delete';
 
+const generateAsyncRequestId = () =>
+  Math.random().toString(36).substring(2, 15) + Math.random().toString(36).substring(2, 15);
+
+const getRequestStatus = (request) => get(request, ['status']);
+const getRequestError = (request) => get(request, ['error']);
+const getRequestAsyncOpId = (request) => get(request, ['data', 'id']);
+
+const isStatusInflight = (status) => status === 'inflight';
+const isStatusSuccess = (status) => status === 'success';
+
 const BulkGranule = ({
   history,
   dispatch,
@@ -25,11 +35,16 @@ const BulkGranule = ({
   const [showModal, setShowModal] = useState(false);
   const [showBulkOpsModal, setShowBulkOpsModal] = useState(false);
   const [showBulkDeleteModal, setShowBulkDeleteModal] = useState(false);
-  const [bulkOpRequestId] = useState(Math.random().toString(36).substring(2, 15) + Math.random().toString(36).substring(2, 15));
-  const [bulkDeleteRequestId] = useState(Math.random().toString(36).substring(2, 15) + Math.random().toString(36).substring(2, 15));
+  const [bulkOpRequestId, setBulkOpRequestId] = useState(generateAsyncRequestId());
+  const [bulkDeleteRequestId, setBulkDeleteRequestId] = useState(generateAsyncRequestId());
 
-  const bulkOperationInfo = get(granules.bulk, [bulkOpRequestId]);
-  const bulkDeleteInfo = get(granules.bulkDelete, [bulkDeleteRequestId]);
+  const bulkOpRequestInfo = get(granules.bulk, [bulkOpRequestId]);
+  const bulkOpRequestStatus = getRequestStatus(bulkOpRequestInfo);
+  const bulkOpRequestError = getRequestError(bulkOpRequestInfo);
+
+  const bulkDeleteRequestInfo = get(granules.bulkDelete, [bulkDeleteRequestId]);
+  const bulkDeleteRequestStatus = getRequestStatus(bulkDeleteRequestInfo);
+  const bulkDeleteRequestError = getRequestError(bulkDeleteRequestInfo);
 
   const ButtonComponent = element;
   const modalClassName = 'bulk_granules';
@@ -62,9 +77,16 @@ const BulkGranule = ({
     setShowBulkOpsModal(true);
   }
 
-  function hideBulkOperationsModal (e) {
+  function hideBulkOperationsModal () {
+    // If last operation succeeded, generate a new request ID so
+    // modal doesn't still show success of last operation
+    if (isStatusSuccess(bulkOpRequestStatus)) {
+      setBulkOpRequestId(generateAsyncRequestId());
+    }
     // clear error from any previous request failure
-    dispatch(bulkGranuleClearError(bulkOpRequestId));
+    if (bulkOpRequestError) {
+      dispatch(bulkGranuleClearError(bulkOpRequestId));
+    }
     setShowBulkOpsModal(false);
   }
 
@@ -74,9 +96,17 @@ const BulkGranule = ({
     setShowBulkDeleteModal(true);
   }
 
-  function hideBulkDeleteModal (e) {
+  function hideBulkDeleteModal () {
+    // If last operation succeeded, generate a new request ID so
+    // modal doesn't still show success of last operation
+    if (isStatusSuccess(bulkDeleteRequestStatus)) {
+      setBulkDeleteRequestId(generateAsyncRequestId());
+    }
     // clear error from any previous request failure
-    dispatch(bulkGranuleDeleteClearError(bulkDeleteRequestId));
+    if (bulkDeleteRequestError) {
+      dispatch(bulkGranuleDeleteClearError(bulkDeleteRequestId));
+    }
+    // clear error from any previous request failure
     setShowBulkDeleteModal(false);
   }
 
@@ -114,26 +144,34 @@ const BulkGranule = ({
         </Modal.Body>
       </Modal>
       <BulkOperationsModal
+        asyncOpId={getRequestAsyncOpId(bulkOpRequestInfo)}
         className={modalClassName}
         dispatch={dispatch}
+        error={bulkOpRequestError}
         showModal={showBulkOpsModal}
         handleSuccessConfirm={handleSuccessConfirm}
+        inflight={isStatusInflight(bulkOpRequestStatus)}
         onCancel={hideBulkOperationsModal}
         onCloseModal={hideBulkOperationsModal}
-        operation={bulkOperationInfo}
+        operation={bulkOpRequestInfo}
         requestId={bulkOpRequestId}
         selected={selected}
+        success={isStatusSuccess(bulkOpRequestStatus)}
       ></BulkOperationsModal>
       <BulkDeleteModal
+        asyncOpId={getRequestAsyncOpId(bulkDeleteRequestInfo)}
         className={modalClassName}
         dispatch={dispatch}
+        error={bulkDeleteRequestError}
         showModal={showBulkDeleteModal}
         handleSuccessConfirm={handleSuccessConfirm}
+        inflight={isStatusInflight(bulkDeleteRequestStatus)}
         onCancel={hideBulkDeleteModal}
         onCloseModal={hideBulkDeleteModal}
-        operation={bulkDeleteInfo}
+        operation={bulkDeleteRequestInfo}
         requestId={bulkDeleteRequestId}
         selected={selected}
+        success={isStatusSuccess(bulkDeleteRequestStatus)}
       ></BulkDeleteModal>
     </>
   );
